DEVELOPMENT/JAMES/requirements-james.txt
pynput
SpeechRecognition
pyobjc-core
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
pyobjc-framework-Vision
